aider: Added bash script in bin to run `yq` using Docker.
# Aider chat conversation:

USER: Is there an official docker for yq?

ASSISTANT: Yes, there is an official Docker image for `yq`. It is maintained by the tool's author and can be found on Docker Hub under the name `mikefarah/yq`. This image allows you to run `yq` in a Docker container, which can be particularly useful if you want to use `yq` in a CI/CD pipeline or if you prefer not to install it directly on your system.

To use the `yq` Docker image, you can pull it from Docker Hub and run it as follows:

```bash
docker pull mikefarah/yq
docker run --rm -v "${PWD}":/workdir mikefarah/yq yq [yq options] [yq command]
```

Replace `[yq options]` and `[yq command]` with the appropriate options and commands for your use case. The `-v` flag mounts the current working directory into the container, which allows `yq` to access files from your host system.

USER: Create a bash script in bin called yq that runs docker (see shellcheck for an example) for yq

ASSISTANT: ```diff
--- /dev/null
+++ bin/yq
@@ ... @@
+#!/bin/bash
+
+# Wrapper script for running yq using Docker
+
+docker run --rm -v "${PWD}":/workdir mikefarah/yq yq "$@"
```