aider: Created a role to load and save version tracking information from a YAML file.
# Aider chat conversation:

USER: What is using the variable git_bug_version command?

ASSISTANT: The variable `git_bug_version_tracking_file` is used in the `vm-setup/roles/git_bug/tasks/main.yml` file to refer to the file that tracks the installed version of `git_bug`. This variable is used in tasks to ensure the version tracking file exists, to slurp its content, and to update the version of `git_bug` after a successful update via the `github_release` role.

However, there is no variable explicitly named `git_bug_version_command` in the provided `tasks/main.yml` file for the `git_bug` role. If you are referring to a command to obtain the current version of `git_bug`, such a command is not shown in the provided content. If you need to implement a task that retrieves the current version of `git_bug` using a command, please provide further details or ask to make the file *read-write* to implement the necessary changes.

USER: Is there a way to setup a pre and post configuration for a playbook?

ASSISTANT: In Ansible, you can set up pre-tasks and post-tasks within your playbook to run configurations before and after your main tasks. These are defined at the same level as the `roles` or `tasks` within a playbook.

Here's an example structure of how you might define pre-tasks and post-tasks in a playbook:

```yaml
- hosts: all
  become: yes
  pre_tasks:
    - name: Run pre-configuration tasks
      # ... define tasks here ...

  roles:
    - role1
    - role2

  tasks:
    - name: Main configuration tasks
      # ... define tasks here ...

  post_tasks:
    - name: Run post-configuration tasks
      # ... define tasks here ...
```

The `pre_tasks` will be executed before any roles or tasks, and `post_tasks` will be executed after roles and tasks. If you need to implement pre-tasks or post-tasks in a specific playbook, please provide the playbook file or ask to add it to the chat for editing.

USER: Can I add a pre and post role?

ASSISTANT: Yes, you can add pre and post roles in an Ansible playbook. However, unlike `pre_tasks` and `post_tasks`, there is no direct `pre_roles` or `post_roles` keyword in Ansible. Instead, you can organize your playbook to execute roles in the desired order, effectively creating pre and post conditions.

Here's an example of how you might structure a playbook to include roles that act as pre and post conditions:

```yaml
- hosts: all
  become: yes
  roles:
    - pre_configuration_role  # This role acts as a pre-role
    - main_configuration_role
    - post_configuration_role # This role acts as a post-role
```

In this structure, `pre_configuration_role` will be executed before `main_configuration_role`, and `post_configuration_role` will be executed after. If you have a specific playbook you want to add pre and post roles to, please provide the playbook file or ask to add it to the chat for editing.

USER: Create a role that both loads the version tracking file into a fact and saves that fact back to the yaml file. It would be called as something like roles: - version_tracking vars: set_facts at the beginning of the playbook and roles: - version_tracking vars: save_facts at the end of the playbook. Unless you have a better way of doing this.

ASSISTANT: To create a new role named `version_tracking` that handles both loading and saving version tracking information, we will need to create a few new files within the role's directory structure. The role will have tasks for setting facts from a YAML file and saving facts back to the YAML file.

Here are the diffs for creating the new role:

1. Create the `tasks/main.yml` file to include sub-tasks based on the provided variable.

```diff
--- /dev/null
+++ vm-setup/roles/version_tracking/tasks/main.yml
@@ ... @@
+---
+# Main tasks for version tracking role
+- name: Load version tracking information
+  include_tasks: load_versions.yml
+  when: version_tracking_action == 'load'
+
+- name: Save version tracking information
+  include_tasks: save_versions.yml
+  when: version_tracking_action == 'save'
```

2. Create the `tasks/load_versions.yml` file to load the version tracking information into a fact.

```diff
--- /dev/null
+++ vm-setup/roles/version_tracking/tasks/load_versions.yml
@@ ... @@
+---
+# Task to load version tracking information from a YAML file
+- name: Slurp version tracking YAML file
+  slurp:
+    src: "{{ version_tracking_file }}"
+  register: version_tracking_content
+
+- name: Set version tracking fact from YAML content
+  set_fact:
+    version_tracking: "{{ (version_tracking_content['content'] | b64decode | from_yaml) }}"
```

3. Create the `tasks/save_versions.yml` file to save the version tracking information back to the YAML file.

```diff
--- /dev/null
+++ vm-setup/roles/version_tracking/tasks/save_versions.yml
@@ ... @@
+---
+# Task to save version tracking information to a YAML file
+- name: Save version tracking fact to YAML file
+  copy:
+    dest: "{{ version_tracking_file }}"
+    content: "{{ version_tracking | to_nice_yaml }}"
+  when: version_tracking is defined
```

4. Create the `defaults/main.yml` file to set default values for the role variables.

```diff
--- /dev/null
+++ vm-setup/roles/version_tracking/defaults/main.yml
@@ ... @@
+---
+# Default variables for version tracking role
+version_tracking_file: "path/to/version_tracking.yml"
+version_tracking_action: "load"  # Default action is to load versions
```

To use this role in a playbook, you would structure it like this:

```yaml
- hosts: all
  roles:
    - role: version_tracking
      vars:
        version_tracking_action: 'load'
        version_tracking_file: 'path/to/your_version_tracking.yml'

  # ... your other roles and tasks ...

  roles:
    - role: version_tracking
      vars:
        version_tracking_action: 'save'
        version_tracking_file: 'path/to/your_version_tracking.yml'
```

This setup allows you to specify when to load and when to save the version tracking information by setting the `version_tracking_action` variable to either `'load'` or `'save'`. The `version_tracking_file` variable should point to the YAML file that contains the version tracking information.