aider: Remove manual version tracking update task in `git_bug` role.
# Aider chat conversation:

USER: I added these *read-write* files: vm-setup/roles/git_bug/tasks/main.yml

ASSISTANT: To update the `git_bug` role to only update the version tracking facts and rely on the playbook to call the `version_tracking` role with the save option, we need to remove the task that manually modifies the version tracking file. Here is the diff for the `vm-setup/roles/git_bug/tasks/main.yml` file:

```diff
--- vm-setup/roles/git_bug/tasks/main.yml
+++ vm-setup/roles/git_bug/tasks/main.yml
@@ -15,10 +15,3 @@
     github_release_asset_pattern: "git-bug_linux_amd64"
     github_release_dest_path: "/usr/local/bin/git-bug"
     github_release_file_mode: "0755"
     github_release_version_mismatch: "{{ git_bug_current_version != git_bug_latest_version }}"

-- name: Update git_bug version in version tracking
-  lineinfile:
-    path: "{{ git_bug_version_tracking_file }}"
-    regexp: '^git_bug:.*$'
-    line: "git_bug: \"{{ github_release_latest_version }}\""
-    create: yes
-    backup: yes
-  when: github_release_version_mismatch and github_release_download_result is changed
```

This diff removes the task that manually updates the version tracking file for `git_bug` in the `git_bug` role. Instead, the role should update the `version_tracking` fact, and the playbook should handle saving this information using the `version_tracking` role with the save option.