aider: Modified debug statements to use `DEBUG_PLAYBOOK | default(false) | bool` structure.
# Aider chat conversation:

USER: If a DEBUG_PLAYBOOK fact doesn't exist and I use when: "{{ DEBUG_PLAYBOOK }}" will ansible error out?

ASSISTANT: Yes, if you use `when: "{{ DEBUG_PLAYBOOK }}"` in an Ansible task and the `DEBUG_PLAYBOOK` fact does not exist, Ansible will error out. This is because Ansible will attempt to evaluate the variable `DEBUG_PLAYBOOK`, and if it is not defined, it will raise an undefined variable error.

To avoid this, you should use the `default` filter to provide a fallback value in case the fact is not defined. For example, you can use `when: DEBUG_PLAYBOOK | default(false) | bool` to ensure that the condition evaluates to `False` if `DEBUG_PLAYBOOK` is not defined, and no error will occur.

Would you like me to make changes to the `install_tflint.yml` file to include this safeguard?

USER: Modify the debug statements to use this structure

ASSISTANT: Here is the diff to modify the debug statements in the `vm-setup/roles/hashicorp/tasks/install_tflint.yml` file to use the structure that checks for the existence of the `DEBUG_PLAYBOOK` fact:

```diff
--- vm-setup/roles/hashicorp/tasks/install_tflint.yml
+++ vm-setup/roles/hashicorp/tasks/install_tflint.yml
@@ -7,7 +7,7 @@
 - name: Dump version information using the github_asset_name variable
   debug:
     msg: "Current version of tflint: {{ hashicorp_asset_current_version }}"
-  when: lookup('env', 'DEBUG_PLAYBOOK') | bool
+  when: DEBUG_PLAYBOOK | default(false) | bool

 - name: Create a temporary file
   tempfile:
@@ -15,7 +15,7 @@
   register: hashicorp_tempfile

 - name: Dump tempfile information
   debug:
     msg: "Tempfile: {{ hashicorp_tempfile }}"
-  when: lookup('env', 'DEBUG_PLAYBOOK') | bool
+  when: DEBUG_PLAYBOOK | default(false) | bool

 - name: Get the new release if there is one
   include_role:
@@ -28,7 +28,7 @@
 - name: Dump version information using the github_asset_name variable
   debug:
     msg: |
       Mismatch: {{ github_release_version_mismatch }}
       Download result: {{ github_release_download_result }}
-  when: lookup('env', 'DEBUG_PLAYBOOK') | bool
+  when: DEBUG_PLAYBOOK | default(false) | bool
```

This diff replaces the `lookup('env', 'DEBUG_PLAYBOOK') | bool` condition with `DEBUG_PLAYBOOK | default(false) | bool` for all debug statements in the file. This ensures that the debug tasks will only run if the `DEBUG_PLAYBOOK` fact is set and evaluates to `True`, and will not cause an error if the fact is not defined.